+++ lib/crawler/worker.ex
@@ -31,6 +31,6 @@ defmodule Crawler.Worker do
     {:noreply, state}
   end
 
-  defp mark_processed(%Page{url: url}), do: Store.processed(url)
-  defp mark_processed(_),               do: nil
+  defp mark_processed({:ok, %Page{url: url}}), do: Store.processed(url)
+  defp mark_processed(_),                      do: nil
 end