+++ test/lib/crawler/scraper_test.exs
@@ -1,0 +1,7 @@
+defmodule Crawler.ScraperTest do
+  use Crawler.TestCase, async: true
+
+  alias Crawler.Scraper
+
+  doctest Scraper
+end